% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateESS.R
\name{approximateESS}
\alias{approximateESS}
\title{Calculates the approximate ESS of Lanfear et al. (2015).}
\usage{
approximateESS(dmat, min.nsamples = (dim(dmat)[1] - 100), alpha = 0.05,
  ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for chains, NOT SQUARED.}

\item{alpha}{The cutoff proportion of the asymptoted used to estimate the average pairwise squared distance.}
}
\description{
Calculates the approximate ESS of Lanfear et al. (2015).
}
\keyword{internal}
