% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_coords.R
\name{as.RFcoords}
\alias{as.RFcoords}
\title{Trees as vectors of splits.}
\usage{
as.RFcoords(trees)
}
\arguments{
\item{trees}{list of trees or multiPhylo object}
}
\value{
Output: a list. $coords is a matrix of trees as coordinates (see details), and $taxa is a list of taxa in each split in the matrix (ordered 1:n).
}
\description{
Transforms sample of trees into their coordinates in (reduced) RF space. Reduction is to the set of coordinates (splits/bipartitions) seen in at least one tree.
}
\details{
In the output$coords, every row is a tree, every column the presence/absence of a split.
}
