% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ASDSF}
\alias{ASDSF}
\title{Calculates the average standard deviation of split frequencies (and other related measures).}
\usage{
ASDSF(x, min.freq = 0.01, summary.fn = mean, weights = "equal")
}
\arguments{
\item{x}{A list. Either a list of multiPhylo objects or a list of split probability vectors.}

\item{min.freq}{Splits less than this frequency (on average) will not be included.}

\item{summary.fn}{Function. The summary of the SDSF. For ASDSF, use mean, for max SDSF, use max.}

\item{weights}{Allows the ASDSF to be weighted. Options are "equal" for standard calculations or "entropy" to weight by split entropy. Only valid for summary.fn=mean}
}
\value{
Output: the ASDSF (or other summary of the SDSF).
}
\description{
Calculates the average standard deviation of split frequencies (and other related measures).
}
