% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_convergence_tests.R
\name{ASDSFPermutationTest}
\alias{ASDSFPermutationTest}
\title{Tests the null hypothesis that a set of chains come from the same distributions against the alternative that they are different.}
\usage{
ASDSFPermutationTest(coords, ngen, nchains, min.split.freq, nrep = 500,
  ...)
}
\arguments{
\item{coords}{A 0/1 matrix of the splits in each tree (RF coordinate matrix).}

\item{ngen}{Number of generations in each chain (cannot mix chain sizes).}

\item{nchains}{Number of chains that are in the distance matrix.}

\item{nrep}{Number of permutation replicates.}
}
\description{
Tests the null hypothesis that a set of chains come from the same distributions against the alternative that they are different.
}
\keyword{internal}
