% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomialProportionCI}
\alias{binomialProportionCI}
\title{Confidence intervals for proportions.}
\usage{
binomialProportionCI(p, n, method, ci.width = 0.95)
}
\arguments{
\item{p}{The (estimated) probability.}

\item{n}{The sample size (or effective sample size).}

\item{method}{Jefreys|Wilson|ContinuityCorrectedWilson.}

\item{ci.width}{Width of the desired confidence interval (value in [0,1]). Equivalent to 1 - 2*alpha.}
}
\value{
A length(p)x2 matrix containing the lower and upper ends of the CI.
}
\description{
Provides three methods for computing confidence intervals for binomial proportions.
Useful for turning estimates of the treESS into CIs for split and topology probabilities.
}
\details{
In practice, the Jeffreys and Wilson intervals often seem to give similar results, while the Corrected Wilson interval often gives much wider intervals.
}
\examples{

p <- rbeta(100,1,1) # Generate probabilities
n <- 100 # Sample size
x <- rbinom(length(p),n,p) # Random variables

# compute CIs
jci <- binomialProportionCI(x/n,n,"jeffreys")
wci <- binomialProportionCI(x/n,n,"wilson")
cci <- binomialProportionCI(x/n,n,"wilson.corrected")

# compare performance
sum(x/n >= jci[,1] & x/n <= jci[,2])/length(p)
sum(x/n >= wci[,1] & x/n <= wci[,2])/length(p)
sum(x/n >= cci[,1] & x/n <= cci[,2])/length(p)
}
\references{
Newcombe (1998). "Two-sided confidence intervals for the single proportion: comparison of seven methods." Statistics in Medicine.
Brown, Cai, and DasGupta (2001). "Interval estimation for a binomial proportion." Statistical science.
}
\seealso{
\link{binomialProportionPI}
}
