% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_probability_utilities.R
\name{binomialProportionDifferenceCI}
\alias{binomialProportionDifferenceCI}
\title{Confidence intervals for differences in proportions.}
\usage{
binomialProportionDifferenceCI(p1, p2, n1, n2, method, ci.width = 0.95)
}
\arguments{
\item{p1}{The (estimated) probability in sample 1.}

\item{p2}{The (estimated) probability in sample 2.}

\item{n1}{The sample size (or effective sample size) in sample 1.}

\item{n2}{The sample size (or effective sample size) in sample 2.}

\item{method}{Either AgrestiCoffo|JeffreysPerks or an option for package DescTool's function BinomDiffCI (if installed).}

\item{ci.width}{Width of the desired confidence interval (value in [0,1]). Equivalent to 1 - 2*alpha.}
}
\value{
A length(p)x2 matrix containing the lower and upper ends of the CI.
}
\description{
Provides two methods for computing confidence intervals for binomial proportions.
Additionally allows for any method available in DescTools::BinomDiffCI.
Useful for turning estimates of the treESS into CIs for differences between split and topology probabilities between chains.
}
\details{
There are a variety of options for differences in binomial proportions with a variety of performance options. DescTools recommends mn, though Agresti-Coffo and Jeffreys-Perks nominally perform well (see references).
}
\examples{

p1 <- rbeta(100,1,1) # Generate probabilities
p2 <- rbeta(100,1,1) # Generate probabilities
n1 <- n2 <- 100 # Sample size
x1 <- rbinom(length(p),n1,p1) # Random variables
x2 <- rbinom(length(p),n2,p2) # Random variables

# compute CIs
jci <- binomialProportionCI(x/n,n,"jeffreys")
wci <- binomialProportionCI(x/n,n,"wilson")
cci <- binomialProportionCI(x/n,n,"wilson.corrected")

# compare performance
sum(x/n >= jci[,1] & x/n <= jci[,2])/length(p)
sum(x/n >= wci[,1] & x/n <= wci[,2])/length(p)
sum(x/n >= cci[,1] & x/n <= cci[,2])/length(p)
}
\references{
Agresti, A., & Caffo, B. (2000). Simple and effective confidence intervals for proportions and differences of proportions result from adding two successes and two failures. The American Statistician, 54(4), 280-288.
Newcombe, R. G. (1998). Interval estimation for the difference between independent proportions: comparison of eleven methods. Statistics in medicine, 17(8), 873-890.
}
\seealso{
\link{binomialProportionPI}, DescTools::BinomDiffCI
}
