% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomialProportionPI}
\alias{binomialProportionPI}
\title{Prediction intervals for proportions.}
\usage{
binomialProportionPI(p, n, pi.width = 0.95, return.counts = FALSE)
}
\arguments{
\item{p}{The (estimated) probability.}

\item{n}{The sample size (or effective sample size).}

\item{pi.width}{Width of the desired prediction interval (value in [0,1]).}
}
\value{
A length(p)x2 matrix containing the lower and upper ends of the prediction interval.
}
\description{
Given a sample size and an estimated proportion, returns interval in which proportions computed from new samples may lie.
Useful for turning estimates of the treESS into ranges of plasible split and topology probabilities from independent MCMC runs.
}
\details{
Method uses a Bayesian approach and is thus most similar to method="Jeffreys" in \link{binomialProportionCI}.
Predictions for proportions are computed assuming the new sample size is the same as the current sample size, and are only valid in that case.
}
\examples{

p <- rbeta(10,1,1) # Generate probabilities
n <- 100 # Sample size
x <- rbinom(length(p),n,p) # Random variables

# compute CIs
binomialProportionPI(x/n,n)
}
\seealso{
\link{binomialProportionCI}
}
