% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteForceMCMCSE.R
\name{bruteForceMCMCSE}
\alias{bruteForceMCMCSE}
\title{Computes quantities needed for calculating Monte Carlo error for trees.}
\usage{
bruteForceMCMCSE(simulated.samples, measures = c("treeProbSquaredError",
  "splitProbSquaredError", "MRCSquaredError"))
}
\arguments{
\item{simulated.samples}{An object of class simulatedPosterior (output of simulatePhylogeneticMCMC).}

\item{measures}{The error or variance measure(s) (see \link{effectiveSizeEquivalentError}).}
}
\value{
A named list with the distribution for each measure. Same as a single element in \link{effectiveSizeEquivalentError}.
}
\description{
Computes quantities needed for calculating Monte Carlo error for trees.
}
\seealso{
\link{effectiveSizeEquivalentError}
}
