% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFrequencyESS.R
\name{cladeFrequencyESS}
\alias{cladeFrequencyESS}
\title{Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).}
\usage{
cladeFrequencyESS(trees)
}
\arguments{
\item{trees}{All trees in the MCMC chain as a multiPhylo object.}
}
\description{
Convenience function wrapper for splitFrequencyESS(x,rooted=TRUE).
}
\references{
Vats and Knudson (2021). "Revisiting the Gelman-Rubin diagnostic." Statistical Science, 36(4), 518-529.
}
\keyword{internal}
