% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMDSESS.R
\name{CMDSESS}
\alias{CMDSESS}
\title{Calculates ESS for a single summary variable obtained from classical multidimensional scaling.}
\usage{
CMDSESS(dmat, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}
}
\description{
Calculates ESS for a single summary variable obtained from classical multidimensional scaling.
}
\keyword{internal}
