% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_convergence_tests.R
\name{CMDSPSRF}
\alias{CMDSPSRF}
\title{Calculates PSRF for a single summary variable obtained from classical multidimensional scaling}
\usage{
CMDSPSRF(dmat, ngen, nchains, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for ALL chains.}

\item{ngen}{Number of generations in each chain (cannot mix chain sizes).}

\item{nchains}{Number of chains that are in the distance matrix.}
}
\description{
Calculates PSRF for a single summary variable obtained from classical multidimensional scaling
}
\keyword{internal}
