% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTrees.R
\name{countTrees}
\alias{countTrees}
\title{Counts occurrences of each tree in a set of trees, returns unique trees and their counts}
\usage{
countTrees(trees, rooted = FALSE)
}
\arguments{
\item{trees}{A multiPhylo object or list of phylo objects}
}
\value{
A list. $trees contains all unique topologies, and $counts their counts
}
\description{
Counts occurrences of each tree in a set of trees, returns unique trees and their counts
}
\examples{
countTrees(rmtree(100,10))
}
