% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakagami.R
\name{dnakagami}
\alias{dnakagami}
\alias{pnakagami}
\alias{qnakagami}
\alias{rnakagami}
\title{Nakagami distribution}
\usage{
dnakagami(x, shape, spread, log = FALSE)

pnakagami(q, shape, spread, lower.tail = TRUE, log.p = FALSE)

qnakagami(p, shape, spread, lower.tail = TRUE, log.p = FALSE)

rnakagami(n, shape, spread)
}
\arguments{
\item{x, }{q vector of quantiles}

\item{shape}{shape parameter > 0.0}

\item{spread}{spread parameter > 0.0}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x), otherwise P(X > x).}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
The Nakagami distribution is the distribution of the square-root of a Gamma-distributed variable.
That is, f X ~ Gamma, then sqrt(X) ~ Nakagami.
}
\details{
It can also be used to approximate the distribution of the standard deviation of split frequencies.


The Nakagami distribution with shape m and scale o has density
f(x) = 2/gamma(m) \* (m/o)^m \* x^(2m - 1) \* exp(-m/o \* x^2)
}
\seealso{
\link{}
}
