% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.mcmc.R
\docType{data}
\name{ds.mcmc}
\alias{ds.mcmc}
\title{MCMC samples of 1000 trees each.}
\format{
A list of ape-style multiPhylo objects (without branch lengths).
}
\source{
Original DS analyses come from Whidden et al. (2020).
}
\usage{
data(ds.mcmc)
}
\description{
Specifically, these are sets of 1000 topologies (there are no branch lengths) produced by using \link{simulatePhylogeneticMCMC} on several standard test datasets collectively referred to as DS1, DS3, and DS5.
These three datasets capture strongly peaked (DS3), flat (DS5), and multimodal (DS1) posterior distributions.
For each, three chains were run resulting in low, medium, and high ESS values
}
\examples{
data(ds.mcmc)
\dontrun{
treess(ds.mcmc,phangorn::RF.dist)
}
}
\references{
Magee, Karcher, Matsen, and Minin (2023). "How trustworthy is your tree? Bayesian phylogenetic effective sample size through the lens of Monte Carlo error." Bayesian Analysis, 1(1), 1-29.

Whidden et al. (2020) "Systematic exploration of the high likelihood set of phylogenetic tree topologies." Systematic Biology 69.2 (2020): 280-293.
}
\keyword{datasets}
