% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveSizeEquivalentError.R
\name{effectiveSizeEquivalentError}
\alias{effectiveSizeEquivalentError}
\title{Estimates distribution of error if we drew ESS trees IID from the (known) posterior distribution.}
\usage{
effectiveSizeEquivalentError(simulated.samples, tree.dist = "RF",
  measures = c("treeProbSquaredError", "splitProbSquaredError",
  "MRCSquaredError"), ess.methods = getESSMethods(), return.ess = TRUE,
  verbose = TRUE)
}
\arguments{
\item{simulated.samples}{An object of class simulatedPosterior (output of simulatePhylogeneticMCMC).}

\item{tree.dist}{The distance measure for trees, only used for ESS computations (RF or SPR, use only one, default RF).}

\item{measures}{The error or variance measure(s) (see details).}

\item{ess.methods}{The ESS calculation method(s) (see details).}

\item{return.ess}{Should the returned lists include the calculated ESS for each chain?}

\item{verbose}{Should progress be printed to screen?}
}
\value{
The first layer are the different ess methods used, the second the performance measures.
So $CMDSESS$distanceToTrueMRC contains the distribution of RF distances to the true MRC tree when drawing ESS samples IID from the true posterior (using CMDSESS for calculating ESS).
}
\description{
Estimates distribution of error if we drew ESS trees IID from the (known) posterior distribution.
}
\details{
There are three options for measuring Monte Carlo error.
1) treeProbSquaredError, yields a vector of squared differences between the per-chain estimate of each topology probability and the average topology probability (averaged over all chains).
2) splitProbSquaredError yields a vector of squared differences between the per-chain estimate of each split probability and the average split probability (averaged over all chains).
3) MRCSquaredError yields a vector of squared RF distances from the MRC of each MCMC chain to the MRC obtained by pooling all chains to compute split frequencies.
}
