% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_simulated_MCMC.R
\name{equivalentError}
\alias{equivalentError}
\title{Obtains estimates of the distribution of error and variance measures from simulated phylogenetic MCMC.}
\usage{
equivalentError(simulated.samples, tree.dist = c("RF", "SPR"),
  measures = c("treeProbSquaredError", "treeVarianceSquaredError",
  "splitProbSquaredError", "distanceToTrueMRC"),
  ess.methods = c("fixedN", getESSMethods()), return.ess = TRUE,
  verbose = TRUE)
}
\arguments{
\item{simulated.samples}{An object of class simulatedPosterior (output of simulatePhylogeneticMCMC).}

\item{tree.dist}{The distance measure for trees (RF or SPR, only one).}

\item{measures}{The error or variance measure(s) (see details).}

\item{ess.methods}{The ESS calculation method(s) (see details).}

\item{return.ess}{Should the returned lists include the calculated ESS for each chain?}

\item{verbose}{Should progress be printed to screen?}
}
\value{
The first layer are the different ess methods used, the second the performance measures.
So $CMDSESS$RFDistanceToTrueMean contains the distribution of RF distances to the distribution of RF distances to the true mean tree when using CMDSESS for calculating ESS.
}
\description{
Obtains estimates of the distribution of error and variance measures from simulated phylogenetic MCMC.
}
\examples{
countTrees(rmtree(100,10))
}
