% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastRF}
\alias{fastRF}
\title{Tree distance}
\usage{
fastRF(trees)
}
\arguments{
\item{trees}{list of trees or multiPhylo object.}
}
\value{
Distances as an object of class dist.
}
\description{
Calculates unrooted RF distance between a set of trees.
}
\details{
Unlike phangorn::RF.dist, does not offer rooted RF distances of warn when unrooting trees.
}
