% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tree.R
\name{find_tree}
\alias{find_tree}
\title{Find the best tree among the trees with the number of nodes at most equal to
the sample size}
\source{
Cand{\`e}s, E. and Tao, T. (2007). The Dantzig selector: Statistical estimation when p is much
larger than n. Annals of Statistics 35 (6), 2313--2351.

Phoa, F. K., Pan, Y. H. and Xu, H. (2009). Analysis of supersaturated
designs via the Dantzig selector. Journal of Statistical Planning and Inference
139 (7), 2362--2372.
}
\usage{
find_tree(MSPEall, leng, n)
}
\arguments{
\item{MSPEall}{a vector with MSPE values for different trees.}

\item{leng}{a vector of the same length as \code{MSPEall} indicating the number of
leaf nodes in each tree.}

\item{n}{the sample size.}
}
\value{
A number indicating the index of the best tree. Note that 1 is returned if
all trees have more than \code{n} nodes.
}
\description{
A function that finds the best tree among all the trees with the number of nodes
at most equal to the sample size
}
\keyword{internal}
