% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finds0Smoothed}
\alias{finds0Smoothed}
\title{Finds s0 for a piecewise constant curve by linear interpolation.}
\usage{
finds0Smoothed(x, y, y.crit)
}
\arguments{
\item{x}{The time lags at which measurements are taken (MUST INCLUDE 0!)}

\item{y}{The measurement associate with each x that we are using for to compute the ESS.}

\item{y.crit}{The critical value above which independence is achieved.}
}
\description{
Curve is assumed to be nondecreasing.
}
\keyword{internal}
