% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0_ess.R
\name{findt0Smoothed}
\alias{findt0Smoothed}
\title{Finds t0 for a piecewise constant curve by linear interpolation.}
\usage{
findt0Smoothed(x, y, y.crit, above.or.below)
}
\arguments{
\item{x}{The time lags at which measurements are taken (MUST INCLUDE 0!)}

\item{y}{The measurement associate with each x that we are using for to compute the ESS.}

\item{y.crit}{The critical value above (or below) which independence is achieved.}

\item{above.or.below}{Is t0 the time at which y >= y.crit ("above") or y <= y.crit ("below")?}
}
\description{
Finds t0 for a piecewise constant curve by linear interpolation.
}
\keyword{internal}
