% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldedRankMedioidESS.R
\name{foldedRankMedioidESS}
\alias{foldedRankMedioidESS}
\title{Calculates an ESS in the spirit of the folded rank-transformed ESS of Vehtari et al. (2019), replacing the median with the medioid.
Where there is no unique medioid, calculates the ESS for all and returns the minimum.}
\usage{
foldedRankMedioidESS(dmat, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}
}
\description{
Calculates an ESS in the spirit of the folded rank-transformed ESS of Vehtari et al. (2019), replacing the median with the medioid.
Where there is no unique medioid, calculates the ESS for all and returns the minimum.
}
\keyword{internal}
