% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{foldedRankMedioidPSRF}
\alias{foldedRankMedioidPSRF}
\title{Calculates an PSRF in the spirit of the folded rank-transformed PSRF of Vehtari et al. (2019), replacing the median with the medioid.
Where there is no unique medioid, calculates the PSRF for all and returns the maximum.}
\usage{
foldedRankMedioidPSRF(dmat, ngen, nchains, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for ALL chains.}

\item{ngen}{Number of generations in each chain (cannot mix chain sizes).}

\item{nchains}{Number of chains that are in the distance matrix.}
}
\description{
Calculates an PSRF in the spirit of the folded rank-transformed PSRF of Vehtari et al. (2019), replacing the median with the medioid.
Where there is no unique medioid, calculates the PSRF for all and returns the maximum.
}
\keyword{internal}
