% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_convergence_tests.R
\name{getConvergenceTestMethods}
\alias{getConvergenceTestMethods}
\title{Lists all available methods for computing test of convergence of multiple MCMC chains of trees/multivariate objects.}
\usage{
getConvergenceTestMethods()
}
\description{
Lists all available methods for computing test of convergence of multiple MCMC chains of trees/multivariate objects.
}
\examples{
getConvergenceTestMethods()
}
\keyword{internal}
