% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treess.R
\name{getESSMethods}
\alias{getESSMethods}
\title{Lists all available methods for computing the ESS of trees/multivariate objects.}
\usage{
getESSMethods(recommended = FALSE)
}
\arguments{
\item{reccomended}{If TRUE, just returns list of the best-performing methods.}
}
\description{
Lists all available methods for computing the ESS of trees/multivariate objects.
}
\examples{
getESSMethods()
}
