% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSparseSpanningTree}
\alias{getSparseSpanningTree}
\title{Gets a sparse representation of a spanning tree from a distance matrix.}
\usage{
getSparseSpanningTree(x)
}
\arguments{
\item{x}{A distance matrix}
}
\value{
Vector of all edges in the spanning tree (contains both i->j and j->i for any connected values i and j)
}
\description{
Gets a sparse representation of a spanning tree from a distance matrix.
}
\keyword{internal}
