% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSparseSpanningTreeListHolmes}
\alias{getSparseSpanningTreeListHolmes}
\title{Gets a sparse representation of a spanning tree from a distance matrix.}
\usage{
getSparseSpanningTreeListHolmes(x, R, labels, se.cutoff = 0.25)
}
\arguments{
\item{x}{A distance matrix}

\item{R}{Either a number of MSTs to collect (to account for non-uniqueness) or "automatic" to chose the number based on the stabilization of the Holmes test statistic}

\item{labels}{The labels, as in holmesTest}

\item{se.cutoff}{For eventual use in auto-determination of number of MSTs needed.}
}
\value{
Vector of all edges in the spanning tree (contains both i->j and j->i for any connected values i and j)
}
\description{
Gets a sparse representation of a spanning tree from a distance matrix.
}
\keyword{internal}
