% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holmesTest.R
\name{holmesTest}
\alias{holmesTest}
\alias{holmesPlot}
\title{Holmes' 2-sample tree test}
\usage{
holmesTest(
  x,
  dist.fn = NULL,
  labels = NULL,
  B = 1000,
  R = 1000,
  returnNullDistribution = FALSE,
  ...
)

holmesPlot(
  x,
  dist.fn = NULL,
  labels = NULL,
  setColors = 1 + (1:min(c(length(x), length(unique(labels))))),
  betweenColor = 1
)
}
\arguments{
\item{x}{Either a list of multiPhylo objects or a distance matrix for distances between trees, if the latter argument "labels" is needed.}

\item{dist.fn}{Function for computing tree to tree distances, must be capable of being called on a multiPhylo. Not needed if x is a distance matrix.}

\item{labels}{If x is a distance matrix, this argument assigns each column and row to a bootstrap replicate.}

\item{B}{Number of bootstrap replicates.}

\item{R}{Number of non-unique MSTs sampled when MST is not unique (see details).}

\item{betweenColor}{Color to distinguish edges between sets for plotting.}

\item{setColor}{Colors for treesets for plotting.}
}
\value{
See details.
}
\description{
Holmes' tree test is a permutation test of equivalence of two (or more) tree distributions.
The input are sets of samples of trees, from which a minimum spanning tree is constructed to link all trees.
The test statistic, S, is the number of edges which connect trees from the same set of samples.
The null distribution is obtained by permuting the labels (the assignment of trees to sets of samples).
}
\details{
holmesTest returns a list with $p.value the (estimated) p-value from the permutation test, $S_0 the observed test statistic, $B the number of bootstrap replicates used, $R the number of replicate MSTs, and optionally $nullDistribution the null distribution of the test statistic obtained by permutation.

When using a topology-only distance, the MST may not be unique (it will certainly not be unique if the same tree occurs multiple times).
In this case, R MSTs are generated and the test statistic used is the average of S across the R trees.
This will greatly increase the computation time required and p-values may nevertheless be unstable (but more stable than without averaging).

holmesPlot returns NULL and plots the visualization.
In the case where the MST is not unique, holmesPlot samples an MST at random and plots the result based on it.
}
\section{Functions}{
\itemize{
\item \code{holmesPlot()}: Uses classical multidimensional scaling to plot trees in a 2D plane and shows edges connecting the minimum spanning tree.

}}
\examples{
\dontrun{
# Get a bunch of trees from two distinct distributions
true1 <- ape::rtree(10,FALSE,br=rexp,rate=170)
true2 <- ape::rtree(10,FALSE,br=rexp,rate=170)

set1 <- lapply(1:100,function(i){
  aln <- phangorn::simSeq(true1,l=200)
  dm <- phangorn::dist.ml(aln)
  ape::bionj(dm)
})
class(set1) <- "multiPhylo"

set2 <- lapply(1:100,function(i){
  aln <- phangorn::simSeq(true2,l=200)
  dm <- phangorn::dist.ml(aln)
  ape::bionj(dm)
})

holmesTest(list(set1,set2),phangorn::RF.dist)
holmesPlot(list(set1,set2),phangorn::RF.dist)
}
}
\references{
Holmes (2005). "Statistical approach to tests involving phylogenies." In Mathematics of Evolution and Phylogeny, ed. Gascuel.
}
