% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holmesTest.R
\name{holmesTestStat}
\alias{holmesTestStat}
\title{Computes test statistic for Holme's tree test from a pre-computed sparse spanning tree and a set of labels}
\usage{
holmesTestStat(sparseSpanningTree, labels)
}
\arguments{
\item{sparseSpanningTree}{The spanning tree as a list of edges}

\item{labels}{Which tree goes in which set.}
}
\value{
The test statistic (number of edges connecting trees within a set)
}
\description{
Computes test statistic for Holme's tree test from a pre-computed sparse spanning tree and a set of labels
}
\keyword{internal}
