% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpDistanceEquivalenceTestESS.R
\name{jumpDistanceBootstrapESS}
\alias{jumpDistanceBootstrapESS}
\title{#' Computes ESS by finding the time lag at which the "jump distance" curve becomes indistinguishable from the jump distances of an uncorrelated set of samples from the same target distribution.
#' Uses permutation resampling to compute the null distribution.
#' Uses linear interpolation to smooth results.
#'
#' @param dmat Sample-to-sample distance matrix for the MCMC chain.
#' @param trees For compatibility with eval and call, not used here.
#' @param nsim Number of simulations (bootstraps/permutations) to compute the null distribution.
#' @param alpha The cutoff quantile for the null distribution, above which samples are considered independent.
#' @param min.nsamples We only consider distances between x_t and x_{t+T} if there are at least this many samples.
#' @param interpolate: Whether to use linear interpolation (default) on the (smoothed) jump distance curve (otherwise it is piecewise constant).
#' @param use.median: Whether to summarize the jump distance distribution at each time lag by the median (if TRUE), otherwise the mean is used.
#' @keywords internal
jumpDistancePermutationESS <- function(dmat,trees=NA,min.nsamples=5,nsim=1000,alpha=0.05) {
  # A shortcut for jumpDistanceEquivalenceTestESS using permutations
  jumpDistanceEquivalenceTestESS(dmat=dmat,min.nsamples=min.nsamples,nsim=nsim,alpha=alpha,bootstrap=FALSE,interpolate=TRUE)
}
Computes ESS by finding the time lag at which the "jump distance" curve becomes indistinguishable from the jump distances of an uncorrelated set of samples from the same target distribution.
Uses (nonparametric) bootstrapping to compute the null distribution.
Uses linear interpolation to smooth results.}
\usage{
jumpDistanceBootstrapESS(dmat, trees = NA, min.nsamples = 5,
  nsim = 1000, alpha = 0.05)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}

\item{trees}{For compatibility with eval and call, not used here.}

\item{min.nsamples}{We only consider distances between x_t and x_{t+T} if there are at least this many samples.}

\item{nsim}{Number of simulations (bootstraps/permutations) to compute the null distribution.}

\item{alpha}{The cutoff quantile for the null distribution, above which samples are considered independent.}

\item{use.median:}{Whether to summarize the jump distance distribution at each time lag by the median (if TRUE), otherwise the mean is used.}
}
\description{
#' Computes ESS by finding the time lag at which the "jump distance" curve becomes indistinguishable from the jump distances of an uncorrelated set of samples from the same target distribution.
#' Uses permutation resampling to compute the null distribution.
#' Uses linear interpolation to smooth results.
#'
#' @param dmat Sample-to-sample distance matrix for the MCMC chain.
#' @param trees For compatibility with eval and call, not used here.
#' @param nsim Number of simulations (bootstraps/permutations) to compute the null distribution.
#' @param alpha The cutoff quantile for the null distribution, above which samples are considered independent.
#' @param min.nsamples We only consider distances between x_t and x_{t+T} if there are at least this many samples.
#' @param interpolate: Whether to use linear interpolation (default) on the (smoothed) jump distance curve (otherwise it is piecewise constant).
#' @param use.median: Whether to summarize the jump distance distribution at each time lag by the median (if TRUE), otherwise the mean is used.
#' @keywords internal
jumpDistancePermutationESS <- function(dmat,trees=NA,min.nsamples=5,nsim=1000,alpha=0.05) {
  # A shortcut for jumpDistanceEquivalenceTestESS using permutations
  jumpDistanceEquivalenceTestESS(dmat=dmat,min.nsamples=min.nsamples,nsim=nsim,alpha=alpha,bootstrap=FALSE,interpolate=TRUE)
}
Computes ESS by finding the time lag at which the "jump distance" curve becomes indistinguishable from the jump distances of an uncorrelated set of samples from the same target distribution.
Uses (nonparametric) bootstrapping to compute the null distribution.
Uses linear interpolation to smooth results.
}
\keyword{internal}
