% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{KL}
\alias{KL}
\title{KL divergence}
\usage{
KL(p, q)
}
\arguments{
\item{p}{probability mass function}

\item{q}{probability mass function}
}
\value{
KL divergence.
}
\description{
Calculates Kullbeck-Leibler divergence KL(p||q) for discrete p and q.
}
