% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{medianDistanceWilcoxTest}
\alias{medianDistanceWilcoxTest}
\title{Computes a Wilcoxon signed rank test on a pair of transformed variables B and W
The original sample is an array X and x_ij is the ith sample of the variable from chain j (x_ij can be a vector, a tree, or any object for which we can compute distances)
Then W_i = median(d(x_ij,x_.j)) and B_i = median(d(x_ij,x_i.))
That is, W is the vector of within-chain median distances (median distance from each point to every other point in the same chain), and B is the vector of between chain median distances
If all chains are sampling the same distribution, then W and B should come from the same distribution
Calculates PSRF for a single summary variable obtained from classical multidimensional scaling}
\usage{
medianDistanceWilcoxTest(dmat, ngen, nchains, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for ALL chains.}

\item{ngen}{Number of generations in each chain (cannot mix chain sizes).}

\item{nchains}{Number of chains that are in the distance matrix.}
}
\description{
Computes a Wilcoxon signed rank test on a pair of transformed variables B and W
The original sample is an array X and x_ij is the ith sample of the variable from chain j (x_ij can be a vector, a tree, or any object for which we can compute distances)
Then W_i = median(d(x_ij,x_.j)) and B_i = median(d(x_ij,x_i.))
That is, W is the vector of within-chain median distances (median distance from each point to every other point in the same chain), and B is the vector of between chain median distances
If all chains are sampling the same distribution, then W and B should come from the same distribution
Calculates PSRF for a single summary variable obtained from classical multidimensional scaling
}
\keyword{internal}
