% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction_ess.R
\name{minPseudoESS}
\alias{minPseudoESS}
\title{Calculates the pseudo-ESS of Lanfear et al. (2015).
Instead of calculating the pseudo-ESS 100 times and reporting summaries, returns the minimum over all possible reference trees.}
\usage{
minPseudoESS(dmat, trees = NA, nsim = NA, alpha = NA,
  min.nsamples = NA)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}

\item{trees}{For compatibility with eval and call, not used here.}

\item{nsim}{For compatibility with eval and call, not used here.}

\item{alpha}{For compatibility with eval and call, not used here.}

\item{min.nsamples}{For compatibility with eval and call, not used here.}
}
\description{
Calculates the pseudo-ESS of Lanfear et al. (2015).
Instead of calculating the pseudo-ESS 100 times and reporting summaries, returns the minimum over all possible reference trees.
}
\keyword{internal}
