% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakagami.R
\name{nakagami.moms}
\alias{nakagami.moms}
\title{Moments of Nakagami distribution}
\usage{
nakagami.moms(shape, spread)
}
\arguments{
\item{shape, }{shape parameter}

\item{spread}{spread parameter}
}
\value{
mean and variance as named numeric vector
}
\description{
Moments of Nakagami distribution
}
\seealso{
\link{dnakagami}
}
