% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakagami.R
\name{nakagami2gamma}
\alias{nakagami2gamma}
\alias{gamma2nakagami}
\title{Conversion between Nakagami and Gamma parameters}
\usage{
nakagami2gamma(shape, spread, return.scale = TRUE)

gamma2nakagami(shape, rate = NA, scale = NA)
}
\arguments{
\item{shape, }{Nakagami or Gamma shape parameters}

\item{spread}{Nakagami spread parameter}

\item{scale, }{rate Gamma scale or rate parameters}
}
\description{
Conversion between Nakagami and Gamma parameters
}
\keyword{internal}
