% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{nSplits}
\alias{nSplits}
\title{Number of unique splits in a collection of \link{perTreeSplits}.}
\usage{
nSplits(tree.splits)
}
\arguments{
\item{tree.splits}{output of perTreeSplits.}
}
\value{
Number of unique splits.
}
\description{
Number of unique splits in a collection of \link{perTreeSplits}.
}
