% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permuteDistanceMatrix.R
\name{permuteDistanceMatrix}
\alias{permuteDistanceMatrix}
\title{Implicitly generate permutations of distance matrices.}
\usage{
permuteDistanceMatrix(dmat, bootstrap = FALSE)
}
\arguments{
\item{dmat}{matrix of distances}

\item{bootstrap}{TRUE/FALSE, if TRUE sampling is done with replacement, if FALSE sampling is done without replacement (permutation only)}
}
\value{
The permuted or bootstrapped distance matrix.
}
\description{
For a distance matrix D = d_ij, generates new distance matrix resulting from permuting or bootstrapping the samples.
Avoids the need to recompute distances when generating many permuted/bootstrapped distance matrices.
}
