% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{perTreeSplits}
\alias{perTreeSplits}
\title{Trees as collections of splits.}
\usage{
perTreeSplits(trees, rooted = FALSE, ref = sort(trees[[1]]$tip.label))
}
\arguments{
\item{trees}{list of trees or multiPhylo object}

\item{rooted}{For rooted trees, if rooted=TRUE, clades are considered rather than splits. Otherwise trees are treated as unrooted.}

\item{ref}{Advanced option. Allows access to ref argument of \link{ape::.compressTipLabel}. Default ensures splits/clades are comparable across calls to different sets of trees on the same taxa, and should only be changed with care!}
}
\value{
Output an object of class perTreeSplits (a list of splits in each tree with attributes for the taxon labels and whether the tree is considered rooted or not).
}
\description{
Transforms sample of trees into their component splits (or clades).
Works like a version of ape::prop.part which returns per-tree (and which considers rooting).
}
\seealso{
\link{splitProbs}, \link{nSplits}, \link{ape::prop.part}
}
