% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_coords.R
\name{perTreeSplits2RFcoords}
\alias{perTreeSplits2RFcoords}
\title{Trees as vectors of splits.}
\usage{
perTreeSplits2RFcoords(splits)
}
\arguments{
\item{splits}{perTreeSplits object}
}
\value{
Output: a list. $coords is a matrix of trees as coordinates (see details), and $taxa is a list of taxa in each split in the matrix (ordered 1:n).
}
\description{
Transforms perTreeSplits list into RF coordinate matrix.
}
\details{
In the output$coords, every row is a tree, every column the presence/absence of a split.
}
\keyword{internal}
