% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseneus_tree_stability.R
\name{plotConsensusTreeStabilityConvergence}
\alias{plotConsensusTreeStabilityConvergence}
\title{Visualizes stability of the consensus tree from a sample of trees over the length of the run.}
\usage{
plotConsensusTreeStabilityConvergence(x, colors = NA,
  CI.color.alpha = 0.5, plot.median = TRUE, ...)
}
\arguments{
\item{x}{Either a multiPhylo object or the output of treeStabilityConvergence.}

\item{colors}{Colors for each of the consensus thresholds for plotting. Any provided transparencies will be removed. If NA defaults are used (defaults not guaranteed to be pretty).}

\item{CI.color.alpha}{Transparency value for plotting CI thresholds}

\item{plot.median}{Is there a median line to be plotted? Only works if 0.5 is included in treeStabilityConvergence(probs=c(...,0.5,...)).}
}
\value{
Nothing, plots the diagnostic.
}
\description{
Visualizes stability of the consensus tree from a sample of trees over the length of the run.
}
\details{
If providing trees, use ... to pass arguments to consensusTreeStabilityConvergence().
}
