% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructTreeIntervals.R
\name{plotTreeIntervals}
\alias{plotTreeIntervals}
\title{Plots intervals for summaries of trees.}
\usage{
plotTreeIntervals(x, summary, chains = c(1, 2), threshold = c(0.5,
  0.75, 0.9), point.col = "black", bar.col = c("springgreen4",
  "firebrick1"), log.axes = NA, xlab = NA, ylab = NA, main = NA)
}
\arguments{
\item{x}{Output of \link{constructTreeIntervals}.}

\item{summary}{"split" to plot split probabilities, "topology" to plot topology probabilities.}

\item{threshold}{If given as a numeric vector, dashed lines are drawn at these values.}

\item{point.col}{Color of points.}

\item{bar.col}{CIs/PIs are plotted using bar.col[1] if the intervals overlap and bar.col[2] otherwise. Default is green and red.}

\item{log.axes}{If TRUE, x- and y-axes are logged. Default is FALSE for splits, TRUE for trees.}

\item{xlab}{The x-axis label, NA for defaults.}

\item{ylab}{The y-axis label, NA for defaults.}

\item{main}{The plot title.}
}
\value{
Nothing, generates plot.
}
\description{
Plots probabilities of splits or trees estimated by two independent MCMC chains, with confidence/prediction intervals.
}
\details{
Colors of cross bars show whether or not the CIs/PIs overlap, highlighting (dis)agreement between chains.
Bars that cross thresholds indicate uncertainty with respect to which side of the threshold a split is on.
By default, thresholds are plotted at p=0.5 (the boundary for being in an MRC tree), 0.75 (moderate support for a split), and 0.9 (relatively strong support for a split).
}
\seealso{
\link{binomialProportionCI}, \link{binomialProportionPI}, \link{treeStability}, \link{constructTreeIntervals}
}
