% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holmesTest.R
\name{prepForHolmes}
\alias{prepForHolmes}
\title{Helper function used to allow multiple input options for holmesPlot and holmesTest}
\usage{
prepForHolmes(x, dist.fn, labels = NULL)
}
\arguments{
\item{x}{Either a list of multiPhylo objects or a distance matrix for distances between trees}

\item{dist.fn}{Function for computing tree to tree distances, must be capable of being called on a multiPhylo.}

\item{labels}{If x is a distance matrix, this argument assigns each column and row to a bootstrap replicate}
}
\value{
Distance matrix
}
\description{
Helper function used to allow multiple input options for holmesPlot and holmesTest
}
\keyword{internal}
