% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoESS.R
\name{pseudoESS}
\alias{pseudoESS}
\title{Calculates the pseudo-ESS of Lanfear et al. (2015).}
\usage{
pseudoESS(dmat, summary = c("min", "median"))
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}

\item{summary}{Should we return the minimum or the median?}
}
\description{
Instead of calculating the pseudo-ESS 100 times and reporting summaries, returns the minimum or median over all possible reference trees at user's discretion.
}
\keyword{internal}
