% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{pseudoPSRF}
\alias{pseudoPSRF}
\title{Calculates the PSRF in the spirit of the pseudo-ESS of Lanfear et al. (2015)
Returns the maximum (over all reference trees in the chain) of all PSRF values.}
\usage{
pseudoPSRF(dmat, ngen, nchains, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for ALL chains.}

\item{ngen}{Number of generations in each chain (cannot mix chain sizes).}

\item{nchains}{Number of chains that are in the distance matrix.}
}
\description{
Calculates the PSRF in the spirit of the pseudo-ESS of Lanfear et al. (2015)
Returns the maximum (over all reference trees in the chain) of all PSRF values.
}
\keyword{internal}
