% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readTreeProbs}
\alias{readTreeProbs}
\title{Reads a MrBayes .trprobs file}
\usage{
readTreeProbs(filepath, CI.width = 1)
}
\arguments{
\item{filepath}{Filepath to MrBayes .trprobs file}

\item{CI.width}{All trees within the posterior CI.width*100 percentile will be returned (1.0 for all trees).}
}
\value{
A list. $trees contains all unique topologies, and $probs their posterior probabilities
}
\description{
Reads a MrBayes .trprobs file
}
