% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_coords.R
\name{RF.from.coords}
\alias{RF.from.coords}
\title{Tree distance}
\usage{
RF.from.coords(coords)
}
\arguments{
\item{coords}{matrix of trees as coordinates, as.RFcoords(trees)$coords.}
}
\value{
Distances as an object of class dist.
}
\description{
Calculates unrooted RF distance between a set of trees.
}
