% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSizesNodes.R
\name{SampleSizesNodes}
\alias{SampleSizesNodes}
\title{Sample sizes in each node}
\usage{
SampleSizesNodes(stratsd, n, stratsize, thhold)
}
\arguments{
\item{stratsd}{a vector specifying the standard deviations of each stratum of the
partition.}

\item{n}{the desired sample size.}

\item{stratsize}{a vector specifying the number of observations in the population
in each stratum of the partition.}

\item{thhold}{a number indicating the minimum number of observations to be selecting
from each stratum.}
}
\value{
A named vector with the same size as the \code{stratsd} and \code{stratsize} with
the number of observations to be selected from each stratum in the sample of size \code{n}.
}
\description{
The function computes the releavant sample sizes in each node of
the given partition.
}
\keyword{internal}
