% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{splitFrechetMean}
\alias{splitFrechetMean}
\title{Split probabilities from collections of splits.}
\usage{
splitFrechetMean(tree.splits)
}
\arguments{
\item{tree.splits}{output of perTreeSplits.}
}
\value{
Named vector of split (or clade) probabilities.
}
\description{
Internal, optionless, error-checking-free version of \link{splitProbs}.
}
\seealso{
\link{splitProbs}, \link{perTreeSplits}
}
\keyword{internal}
