% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{splitFrechetVariance}
\alias{splitFrechetVariance}
\title{Split Frechet variance from collections of splits.}
\usage{
splitFrechetVariance(tree.splits, mean = NULL, bessel = TRUE)
}
\arguments{
\item{tree.splits}{output of perTreeSplits.}

\item{mean}{output of splitFrechetMean}

\item{bessel}{Should Bessel correction (divide by n-1 not n) be used? Default TRUE}
}
\value{
Named vector of split (or clade) probabilities.
}
\description{
Split Frechet variance from collections of splits.
}
\details{
This function interchanges the order of summation.
The Frechet variance is the (average) squared deviation from each split vector to the mean.
Assuming all splits are collected into an nTrees x nSplits matrix X, this is sum_i sum_j (x_ij - mu_j)^2.
By interchanging the order of summation, the problem becomes a sum of variances of vectors of Bernoulli trials.
Thus, the Frechet split variance can be obtained directly from the mean and the number of trees.
}
\seealso{
\link{splitProbs}, \link{perTreeSplits}, \link{splitFrechetMean}.
}
\keyword{internal}
