% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFrequencyESS.R
\name{splitFrequencyESS}
\alias{splitFrequencyESS}
\title{Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).}
\usage{
splitFrequencyESS(trees, ...)
}
\arguments{
\item{trees}{All trees in the MCMC chain, either as trees or as RF coordinates.}
}
\description{
Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).
}
\references{
Vats and Knudson (2021). "Revisiting the Gelman-Rubin diagnostic." Statistical Science, 36(4), 518-529.
}
\keyword{internal}
