% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFrequencyESS.R
\name{splitFrequencyESSold}
\alias{splitFrequencyESSold}
\title{Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).}
\usage{
splitFrequencyESSold(trees, ...)
}
\arguments{
\item{trees}{All trees in the MCMC chain, either as trees or as RF coordinates.}
}
\description{
Older version of function.
Uses internally a matrix-based representation of trees which takes lots of memory for large trees.
Not recommended for large trees.
}
\references{
Vats and Knudson (2021). "Revisiting the Gelman-Rubin diagnostic." Statistical Science, 36(4), 518-529.
}
\keyword{internal}
