% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitProbLikelihoodRatioTest.R
\name{splitProbLikelihoodRatioTest}
\alias{splitProbLikelihoodRatioTest}
\title{Examine differences in split probabilities among chains}
\usage{
splitProbLikelihoodRatioTest(x, ESS, rooted = FALSE, log.p = FALSE)
}
\arguments{
\item{x}{A list of: trees, perTreeSplits, or split probability vectors.}

\item{ESS}{The effective sample sizes for the MCMC chains in x, or a target ESS cutoff as in Fabreti and Höhna (2022) (see details).}

\item{rooted}{If trees are being fed into x, should splits (FALSE) or clades (TRUE) be retrieved.}

\item{log.p}{If TRUE and return.p.value == TRUE, log(p) is returned instead of p.}
}
\value{
The p-values from a LRT for differences in proportion for each observed split.
}
\description{
Takes split (or clade) probabilities from multiple MCMC runs and compares them using a likelihood ratio test, possibly accounting for differing effective sample sizes among chains.
}
\details{
For a given split (or clade), a LRT is conducted testing the null hypothesis that the probability of the split is the same in all chains.
(Splits are tested independently.)
Instead of using the raw number of times a split is seen in a particular chain, an adjusted count is used to accommodate Monte Carlo error.
Specifically, for a split with observed frequency p in a chain with ESS effective samples of the phylogeny, the adjusted count used is p * ESS.

If the actual effective sample size is used, this test generalizes the confidence-interval based approach used in \link{binomialProportionDifferenceCI} to >2 chains.

Alternately, a fixed sample size can be fed in, such as the 625 suggested by Fabreti and Höhna (2022).
In this case, the acceptable difference in split frequencies is fixed, and as the chains are run longer (targeting the same mode) eventually all p-values will exceed a given cutoff.
This is similar to the approach used in convenience::plotDiffSplits.
}
\examples{
\dontrun{
# Load data, compute ESS
data(ds.mcmc)
chains <- list(ds.mcmc$ds3.low,ds.mcmc$ds3.med,ds.mcmc$ds3.high)
ess <- treess(chains,phangorn::RF.dist,methods="frechetCorrelationESS")

# Examine differences in split probabilities accounting for Monte Carlo error
pvals.ess <- splitProbLikelihoodRatioTest(chains,as.numeric(ess))
hist(pvals.ess,breaks=seq(0,1,0.05),xlab="p-value",main="Comparison of split probabilities in DS3")

# Examine differences in split probabilities more like convenience::plotDiffSplits
alpha <- 0.05 # a p-value threshold
pvals.fixed <- splitProbLikelihoodRatioTest(chains,625,log.p=TRUE)
split.probs <- splitProbs(perTreeSplits(do.call(c,chains)))
key <- match(names(split.probs),names(pvals.fixed))
plot(split.probs,-pvals.fixed[key],xlab="split frequency",ylab="-log(LRT p-value)",main="incongruent splits are those above the red line")
abline(h=-log(alpha),col="red")
}
}
\references{
Fabreti and Höhna (2022). "Convergence assessment for Bayesian phylogenetic analysis using MCMC simulation." Methods in Ecology and Evolution, 13(1), 77-90.

Magee, Karcher, Matsen, and Minin (2023). "How trustworthy is your tree? Bayesian phylogenetic effective sample size through the lens of Monte Carlo error." Bayesian Analysis, 1(1), 1-29.
}
\seealso{
\link{compareChainProbabilities}, \link{binomialProportionDifferenceCI}, \link{ASDSF}
}
