% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{splitProbs}
\alias{splitProbs}
\title{Split probabilities from collections of splits.}
\usage{
splitProbs(tree.splits, range = NULL)
}
\arguments{
\item{tree.splits}{output of perTreeSplits.}

\item{range}{optional argument to define which trees to compute probabilities for, defaults to all trees.}

\item{return.names}{should we return the taxa names?}
}
\value{
Named vector of split (or clade) probabilities.
}
\description{
Computes split probabilities from output of \link{perTreeSplits}.
Works like ape::prop.clade to perTreeSplit's ape::prop.part.
}
\seealso{
\link{perTreeSplits}
}
