% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_tests.R
\name{testConvergence}
\alias{testConvergence}
\title{Test for global convergence of multiple MCMC chains.}
\usage{
testConvergence(x, dist.fn, methods = getConvergenceTestMethods(),
  alpha = 0.05, threshold = 1.01, identify.outliers = TRUE,
  min.split.freq = 0.01)
}
\arguments{
\item{x}{List of all chains of the tree-valued (or multivariate or non-Euclidean) variable(s) for which to compute convergence.}

\item{dist.fn}{A function suitable to calculate distances between all values in x.}

\item{alpha}{For methods using  (only matters when simplify.output == TRUE).}

\item{threshold}{A maximum threshold for PSRF to de}

\item{identify.outliers}{If TRUE, and global convergence is not found (for any method), identifies which chain(s) are different from the others. Only works if there are at least 3 chains}

\item{min.split.freq}{For ASDSF-based tests only. Minimum split frequency to be included in ASDSF.}

\item{method}{The method for computing the convergence diagnostic (see details).}
}
\value{
A list, where $converged is a logical vector indicating whether global convergence was found for each method used, $statistic is the computed PSRF or test statistic for each method, and (optionally) $pairwise is a list of matrices with the results of all pairs of tests for all methods.
}
\description{
Test for global convergence of multiple MCMC chains.
}
\details{
Allowable methods are "pseudoPSRF", "totalDistancePSRF", "foldedRankMedioidPSRF", and "medianDistanceWilcoxTest"
}
\examples{
testConvergence(list(rnorm(100),rnorm(100)))
testConvergence(list(rnorm(100,0,1),list(rnorm(100,0,1),list(rnorm(100,0,1),rnorm(100,1,1)))
}
