% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalDistanceESS.R
\name{totalDistanceESS}
\alias{totalDistanceESS}
\title{Calculates the ESS of a transformed variable y where y[i] = sum_j(d(x[i],x[j])).}
\usage{
totalDistanceESS(dmat, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}

\item{trees}{For compatibility with eval and call, not used here.}

\item{nsim}{For compatibility with eval and call, not used here.}

\item{alpha}{For compatibility with eval and call, not used here.}

\item{min.nsamples}{For compatibility with eval and call, not used here.}
}
\description{
Calculates the ESS of a transformed variable y where y[i] = sum_j(d(x[i],x[j])).
}
\keyword{internal}
