% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeStability.R
\name{treeStability}
\alias{treeStability}
\title{Examines stability of tree-based summaries from a sample of trees using a block-bootstrap to resample the MCMC run.}
\usage{
treeStability(trees, consensus.threshold = c(0.5, 0.75, 0.95),
  nrep = 100, min.split.freq = 0.01, probs = c(0.05, 0.5, 0.95),
  ACT = "sqrt")
}
\arguments{
\item{trees}{The sample of trees from the posterior. A multiPhylo object, or a coordinate matrix as from as.RFcoords(trees)$coords.}

\item{consensus.threshold}{The consensus tree will contain only splits above this probability (p > 0.5). Can be a vector of probabilities.}

\item{nrep}{The number of bootstrap replicates used to assess the tree stability.}

\item{min.split.freq}{For ASDSF only. Minimum split frequency to be included in ASDSF.}

\item{probs}{The quantiles in the results summary.}

\item{ACT}{Autocorrelation time for computing the block size for bootstrapping the chain. Default ("sqrt") conservatively scales ACT with sqrt of subsample length. Can also be provided as numeric.}
}
\value{
One element of the list returned by \link{treeStabilityConvergence}.
}
\description{
This is a convenience wrapper for treeStabilityConvergence.
It allows you to compute uncertainty about tree summaries for the full chain and only the full chain.
For more, see \link{treeStabilityConvergence}.
}
\seealso{
\link{treeStabilityConvergence},\link{plotTreeStabilityConvergence}
}
