% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseneus_tree_stability.R
\name{treeStabilityConvergence}
\alias{treeStabilityConvergence}
\title{For some split probability threshold(s), examines stability of the consensus tree from a sample of trees over the length of the run.}
\usage{
treeStabilityConvergence(trees, stat = "MRC", sizes = 10,
  consensus.threshold = c(0.5, 0.75, 0.95), nrep = 100,
  probs = c(0.05, 0.5, 0.95), spacing = "logarithmic",
  min.split.freq = 0.01, ...)
}
\arguments{
\item{trees}{The sample of trees from the posterior. A multiPhylo object, or a coordinate matrix as from as.RFcoords(trees)$coords (only usable with stat="ASDSF|MRC")}

\item{stat}{The summary to be plotted. ASDSF vizualizes convergence of split frequencies, MRC of the summary tree(s), topoProbs of the tree probabilities, imbalance of the tree (im)balance.}

\item{sizes}{Either an integer giving the number of points to compute the stability (at least 3 and no more than the number of trees), or the sample sizes at which the stability should be computed.}

\item{consensus.threshold}{The consensus tree will contain only splits above this probability (p >= 0.5). Can be a vector of probabilities.}

\item{nrep}{The number of (bootstrap) replicates used to assess the consensus tree stability.}

\item{probs}{The quantiles in the results summary.}

\item{spacing}{Options "linear|logarithmic". When determining the sequence of sizes (sizes given as an integer), should they be linearly or logarithmically spaced?}

\item{min.split.freq}{For ASDSF only. Minimum split frequency to be included in ASDSF.}
}
\value{
An object of class conTreeStability (a list). For each sample size, $quantiles contains the quantiles specified by probs for all thresholds. $boot contains a matrix of all the bootstrapped distances.
}
\description{
For some split probability threshold(s), examines stability of the consensus tree from a sample of trees over the length of the run.
}
\details{
Stability is assessed as the RF distance from (bootstrap) replicate consensus trees to the consensus tree from all samples. The real sample sizes taken will not be exactly those provided because of the block sizes for bootstrapping.
}
