% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unrootedImbalance.R
\name{unrootedImbalance}
\alias{unrootedImbalance}
\title{Computes the Colless imbalance measure marginalized over all rootings of the tree.}
\usage{
unrootedImbalance(phy)
}
\arguments{
\item{phy}{The tree as a class phylo object.}
}
\value{
The imbalance measure.
}
\description{
Computes the Colless imbalance measure marginalized over all rootings of the tree.
}
