% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{weightedSplitFrechetMean}
\alias{weightedSplitFrechetMean}
\title{Split Frechet variance from collections of splits which came from tree samples with uneven lengths.}
\usage{
weightedSplitFrechetMean(split.prob.list, w)
}
\arguments{
\item{split.prob.list}{List of outputs of perTreeSplits.}

\item{w}{Weights (e.g., lengths of each set of trees)}
}
\value{
Named vector of split (or clade) probabilities.
}
\description{
For tree samples of even length, the unweighted mean is faster and should be used.
}
\keyword{internal}
