% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_probability_utilities.R
\name{binomialProportionPI}
\alias{binomialProportionPI}
\title{Prediction intervals for proportions.}
\usage{
binomialProportionPI(p, n, n.new = n, pi.width = 0.95)
}
\arguments{
\item{p}{The (estimated) probability.}

\item{n}{The sample size (or effective sample size) from which the probabilities were computed.}

\item{n.new}{The sample size (or effective sample size) of the new dataset for which prediction intervals are to be computed.}

\item{pi.width}{Width of the desired prediction interval (value in [0,1]). Equivalent to 1 - 2*alpha.}
}
\value{
A length(p)x2 matrix containing the lower and upper ends of the prediction interval.
}
\description{
Given a sample size and an estimated proportion, returns interval in which proportions computed from new samples may lie.
Useful for turning estimates of the treESS into ranges of plasible split and topology probabilities from independent MCMC runs.
}
\details{
Method uses a Bayesian approach and is thus most similar to method="Jeffreys" in \link{binomialProportionCI}.
The method returns probabilities from a new sample, counts divided by the sample size.
}
\examples{

p <- rbeta(10,1,1) # Generate probabilities
n <- 100 # Sample size
x <- rbinom(length(p),n,p) # Random variables

# compute CIs
binomialProportionPI(x/n,n)
}
\seealso{
\link{binomialProportionCI}
}
