% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_simulated_MCMC.R
\name{bruteForceMCMCSE}
\alias{bruteForceMCMCSE}
\title{Obtains estimates of the distribution of error and variance measures from simulated phylogenetic MCMC.}
\usage{
bruteForceMCMCSE(simulated.samples, tree.dist = c("RF", "SPR"),
  measures = c("treeProbSquaredError", "treeVarianceSquaredError",
  "splitProbSquaredError", "distanceToTrueMRC"))
}
\arguments{
\item{simulated.samples}{An object of class simulatedPosterior (output of simulatePhylogeneticMCMC).}

\item{tree.dist}{The distance measure for trees (RF or SPR, only one).}

\item{measures}{The error or variance measure(s) (see details).}
}
\value{
A named list with the distribution for each measure.
}
\description{
Obtains estimates of the distribution of error and variance measures from simulated phylogenetic MCMC.
}
\examples{
countTrees(rmtree(100,10))
}
