% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holmesTest.R
\name{holmesTestStat}
\alias{holmesTestStat}
\title{Computes test statistic for Holme's tree test from a pre-computed sparse spanning tree and a set of labels}
\usage{
holmesTestStat(sparseSpanningTrees, labels)
}
\arguments{
\item{labels}{Which tree goes in which set.}

\item{sparseSpanningTree}{A list of spanning trees (each provided as a list of edges)}
}
\value{
The test statistic (number of edges connecting trees within a set)
}
\description{
Allows a list of MSTs, for situations where the MST is not unique, in which case it averages the test statistic over these.
}
\keyword{internal}
