% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoESS.R
\name{medianPseudoESS}
\alias{medianPseudoESS}
\title{Calculates the pseudo-ESS of Lanfear et al. (2015).}
\usage{
medianPseudoESS(dmat, ...)
}
\arguments{
\item{dmat}{Sample-to-sample distance matrix for the MCMC chain.}
}
\description{
Instead of calculating the pseudo-ESS 100 times and reporting summaries, returns the median over all possible reference trees.
}
\keyword{internal}
