% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSplits.R
\name{perTreeSplitsConvergenceHelper}
\alias{perTreeSplitsConvergenceHelper}
\title{Useful quantities from collections of split Frechet means.}
\usage{
perTreeSplitsConvergenceHelper(
  mean.list,
  weights = 1,
  return.ssq = FALSE,
  return.sdsf = FALSE
)
}
\arguments{
\item{mean.list}{list of output of \link{splitFrechetMean} or \link{splitProbs}.}
}
\value{
List with grand mean ($mean), and optionally sum of squared deviations to the grand mean ($ssq) 
and optionally the standard deviations of all split frequencies.
}
\description{
Computes the grand mean from a collection of split Frechet means.
Can also return the sum of squared deviations to that grand mean and standard deviations of split frequencies.
Assumes equal weighting!
}
\seealso{
\link{splitProbs}, \link{perTreeSplits}
}
\keyword{internal}
