% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.circular_scan}
\alias{print.circular_scan}
\title{Print Method for circular_scan Objects}
\usage{
\method{print}{circular_scan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"circular_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for circular_scan Objects
}
