% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tree_scan}
\alias{print.tree_scan}
\title{Print Method for tree_scan Objects}
\usage{
\method{print}{tree_scan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tree_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for tree_scan Objects
}
