% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.treespatial_scan}
\alias{print.treespatial_scan}
\title{Print Method for treespatial_scan Objects}
\usage{
\method{print}{treespatial_scan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"treespatial_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for treespatial_scan Objects
}
