% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogeneticMCMC.R
\name{simulatePhylogeneticMCMC}
\alias{simulatePhylogeneticMCMC}
\title{Counts occurences of each tree in a set of trees, returns unique trees and their counts}
\usage{
simulatePhylogeneticMCMC(adjacency.graph, ngen = 1000, nchains = 100,
  thin = 1, move.weights = c(2, 1), verbose = TRUE)
}
\arguments{
\item{adjacency.graph}{An object of class treeAdjacencyGraph.}

\item{ngen}{Number of MCMC generations to run.}

\item{nchains}{Number of (independent) MCMC chains to run.}

\item{thin}{Samples will be thinned to one every thin generations.}

\item{move.weights}{Relative probability at each step of each move type (NNI,SPR).}

\item{verbose}{Whether to print summaries of MCMC performance.}
}
\value{
Object of class simulatedPosterior, a list. 
$trees contains the unique trees in the target posterior (adjacency.graph$trees).
$probs contains the probabilities of the trees in the target posterior (adjacency.graph$probs).
$indices contains a matrix of the MCMC chains as the indices of the trees at each generation, chains in columns, generations in rows.
}
\description{
Counts occurences of each tree in a set of trees, returns unique trees and their counts
}
\details{
No burnin is required, the chain is initialized from the stationary distribution.
}
\examples{
data(DS3adj)
post <- simulatePhylogeneticMCMC(DS3adj,ngen=50,nchains=1)
post$trees[post$indices[,1]]
}
