% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_by_generalizations.R
\name{splitFrequencyESS}
\alias{splitFrequencyESS}
\title{Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).}
\usage{
splitFrequencyESS(dmat = NA, trees, min.nsamples = NA, alpha = NA,
  nsim = NA)
}
\arguments{
\item{dmat}{For compatibility with eval and call, not used here.}

\item{trees}{All trees in the MCMC chain.}

\item{min.nsamples}{For compatibility with eval and call, not used here.}

\item{alpha}{For compatibility with eval and call, not used here.}

\item{nsim}{For compatibility with eval and call, not used here.}
}
\description{
Calculates ESS using a Frechet-like generalization of the univariate ESS of Vats and Knudson (2018).
}
\references{
Vats, Dootika, and Christina Knudson. "Revisiting the Gelman-Rubin diagnostic." arXiv preprint arXiv:1812.09384 (2018).
}
\keyword{internal}
