% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.circular_scan}
\alias{summary.circular_scan}
\title{Summary Method for circular_scan Objects}
\usage{
\method{summary}{circular_scan}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"circular_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Method for circular_scan Objects
}
