% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.tree_scan}
\alias{summary.tree_scan}
\title{Summary Method for tree_scan Objects}
\usage{
\method{summary}{tree_scan}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tree_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Method for tree_scan Objects
}
