% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.treespatial_scan}
\alias{summary.treespatial_scan}
\title{Summary Method for treespatial_scan Objects}
\usage{
\method{summary}{treespatial_scan}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"treespatial_scan"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary Method for treespatial_scan Objects
}
