% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treess.R
\name{treess}
\alias{treess}
\title{Calculate one or more effective sample size measures for an MCMC log file containing phylogenetic trees or other multivariate/non-Euclidean objects.}
\usage{
treess(x, dist.fn, methods = getESSMethods(recommended = TRUE),
  alpha = 0.05, nsim = 1000, min.nsamples = 5,
  max.approximateESS.timelag = 100)
}
\arguments{
\item{x}{A list of chains containing tree-valued (or multivariate or non-Euclidean) variable for which to compute ESS.}

\item{dist.fn}{A function suitable to calculate distances between all values in x.}

\item{alpha}{Type I error rate for methods using CIs/hypothesis tests, the proportion of the asymptote used in the approximateESS.}

\item{nsim}{For methods using permutation/bootstrap resampling, the number of resampling iterations.}

\item{min.nsamples}{The minimum number of samples do be used in calculating summaries (median distance, correlation, etc.). Essentially the maximum time lag considered is length(chains[[i]]) - min.nsamples. Not applicable to dimension reduction methods.}

\item{max.approximateESS.timelag}{The maximum time lag considered in the approximateESS, overrides min.nsamples.}

\item{method}{The method for computing ESS (see details).}
}
\description{
Calculate one or more effective sample size measures for an MCMC log file containing phylogenetic trees or other multivariate/non-Euclidean objects.
}
\examples{
chains <- list(rnorm(100),cumsum(rnorm(100)))
treess(chains,dist)
}
