% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeStability.R
\name{treeStabilityConvergence}
\alias{treeStabilityConvergence}
\title{Examines stability of tree-based summaries from a sample of trees using a block-bootstrap to resample the MCMC run.
By performing this resampling at a variety of chain lengths, provides information about how estimates have converged over the run.
Summaries include ASDSF, tree imbalance, tree probabilities, and the MRC tree.
This function can be slow with large samples of trees due to the initial cost of computing the distance matrix of all trees.}
\usage{
treeStabilityConvergence(trees, stat = "MRC", sizes = 10,
  consensus.threshold = c(0.5, 0.75, 0.95), nrep = 100,
  probs = c(0.05, 0.5, 0.95), spacing = "logarithmic",
  min.split.freq = 0.01, block.size = "sqrt", ...)
}
\arguments{
\item{trees}{The sample of trees from the posterior. A multiPhylo object, or a coordinate matrix as from as.RFcoords(trees)$coords (only usable with stat="ASDSF|MRC")}

\item{stat}{The summary to be plotted. ASDSF vizualizes convergence of split frequencies, MRC of the summary tree(s), topoProbs of the tree probabilities, imbalance of the tree (im)balance.}

\item{sizes}{Either an integer giving the number of points to compute the stability (at least 3 and no more than the number of trees), or the sample sizes at which the stability should be computed.}

\item{consensus.threshold}{The consensus tree will contain only splits above this probability (p > 0.5). Can be a vector of probabilities.}

\item{nrep}{The number of (bootstrap) replicates used to assess the consensus tree stability.}

\item{probs}{The quantiles in the results summary.}

\item{spacing}{Options "linear|logarithmic". When determining the sequence of sizes (sizes given as an integer), should they be linearly or logarithmically spaced?}

\item{min.split.freq}{For ASDSF only. Minimum split frequency to be included in ASDSF.}

\item{block.size}{Determines the block size for block bootstrapping. Option "sqrt" scales the block size as the square root of the chain length. Otherwise an integer must be specified, this is not advised unless only evaluating at the full chain.}
}
\value{
An object of class conTreeStability (a list). For each sample size, $quantiles contains the quantiles specified by probs for all thresholds. $boot contains a matrix of all the bootstrapped distances.
}
\description{
Examines stability of tree-based summaries from a sample of trees using a block-bootstrap to resample the MCMC run.
By performing this resampling at a variety of chain lengths, provides information about how estimates have converged over the run.
Summaries include ASDSF, tree imbalance, tree probabilities, and the MRC tree.
This function can be slow with large samples of trees due to the initial cost of computing the distance matrix of all trees.
}
\details{
Stability is assessed as the RF distance from (bootstrap) replicate consensus trees to the consensus tree from all samples. The real sample sizes taken will not be exactly those provided because of the block sizes for bootstrapping.
}
